+++ src/modules/index.ts
@@ -1,1 +1,2 @@
 export * from './grid.module';
+export * from './matrixGrid.module';