+++ src/plugins/contentRenderer/matrix/misc/components/tableDefaultTemplates/tableDefaultTemplates.component.ts
@@ -11,6 +11,12 @@ import {HeaderContainerSAComponent} from '../../../../../../components/headerCon
 import {FooterContainerSAComponent} from '../../../../../../components/footerContainer/footerContainer.component';
 import {HeaderContainerTemplateSADirective} from '../../../../../../directives/headerContainerTemplate/headerContainerTemplate.directive';
 import {FooterContainerTemplateSADirective} from '../../../../../../directives/footerContainerTemplate/footerContainerTemplate.directive';
+import {HeaderRowContainerSAComponent} from '../../../../../../components/headerRowContainer/headerRowContainer.component';
+import {ContentRowContainerSAComponent} from '../../../../../../components/contentRowContainer/contentRowContainer.component';
+import {FooterRowContainerSAComponent} from '../../../../../../components/footerRowContainer/footerRowContainer.component';
+import {HeaderRowContainerTemplateSADirective} from '../../../../../../directives/headerRowContainerTemplate/headerRowContainerTemplate.directive';
+import {ContentRowContainerTemplateSADirective} from '../../../../../../directives/contentRowContainerTemplate/contentRowContainerTemplate.directive';
+import {FooterRowContainerTemplateSADirective} from '../../../../../../directives/footerRowContainerTemplate/footerRowContainerTemplate.directive';
 
 /**
  * Component that stores default templates for table content renderig
@@ -27,10 +33,16 @@ import {FooterContainerTemplateSADirective} from '../../../../../../directives/f
         HeaderContainerSAComponent,
         ContentContainerSAComponent,
         FooterContainerSAComponent,
+        HeaderRowContainerSAComponent,
+        ContentRowContainerSAComponent,
+        FooterRowContainerSAComponent,
         GridContainerTemplateSADirective,
         HeaderContainerTemplateSADirective,
         ContentContainerTemplateSADirective,
         FooterContainerTemplateSADirective,
+        HeaderRowContainerTemplateSADirective,
+        ContentRowContainerTemplateSADirective,
+        FooterRowContainerTemplateSADirective,
     ],
     changeDetection: ChangeDetectionStrategy.OnPush
 })